list, detail에 작성자 정보가 나오고 게시물 수정, 삭제 권한체크, 로그인체크. jsp include 사용법

+++ src/main/webapp/jsp/article/detail.jsp
@@ -4,8 +4,11 @@
     pageEncoding="UTF-8"%>
     
 <%
-Map<String, Object> articleRow = (Map) request.getAttribute("articleRow");
-%>
+	Map<String, Object> articleRow = (Map) request.getAttribute("articleRow");
+	int loginedMemberId = (int) request.getAttribute("loginedMemberId");
+%>  
+
+
 
 <!DOCTYPE html>
 <html>
@@ -18,12 +21,21 @@ Map<String, Object> articleRow = (Map) request.getAttribute("articleRow");
 	
 	<div>번호 : <%= (int) articleRow.get("id") %></div>
 	<div>날짜 : <%= (LocalDateTime) articleRow.get("regDate") %></div>
+	<div>작성자 : <%= (String) articleRow.get("writerName") %></div>
 	<div>제목 : <%= (String) articleRow.get("title") %></div>
 	<div>내용 : <%= (String) articleRow.get("body") %></div>
 	<div>
 		<a href="list">목록</a>
-		<a href="modify?id=<%= (int) articleRow.get("id") %>">수정</a>
-		<a href="doDelete?id=<%= (int) articleRow.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+		<%
+		if (loginedMemberId != -1 && (int) articleRow.get("memberId") == loginedMemberId) {
+		%>
+			<a href="modify?id=<%= (int) articleRow.get("id") %>">수정</a>
+			<a href="doDelete?id=<%= (int) articleRow.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+		<%
+		}
+		%>
+
+		
 	</div>
 </body>
 </html>
